Backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
Werkzeug==3.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
